Error handling, refactoring

+++ wwwroot/src/components/ContentTypes.tsx
@@ -1,19 +1,23 @@
-import React, { Suspense } from "react";
+import React from "react";
 import { Dispatch } from "redux";
 import { connect } from "react-redux";
-import { action, LOAD, DELETE } from "../actions/contentTypes";
+import { 
+  action, 
+  CONTENT_TYPE_LIST_LOAD, 
+  CONTENT_TYPE_DELETE, 
+  CONTENT_TYPE_ERROR } from "../actions/contentTypes";
 import { IContentType, IContentTypesState } from "../interfaces";
 
 class ContentTypes extends React.Component<any, IContentTypesState> {
   componentDidMount() {
     this.props.load();
-  }
-  delete_click(contentTypeID: Number): void {
-    this.props.delete(contentTypeID);
-  }
+  }  
   render() {
     return (
-      <Suspense fallback={<div>Loading...</div>}>
+      <>
+        {this.props.error != null
+          ? <div className="error">Ошибка: {this.props.error}</div> 
+          : ""}
         <a href="/komandir/contentTypes/0">Create new</a>
         <ul>
           {this.props.contentTypes &&
@@ -27,9 +31,7 @@ class ContentTypes extends React.Component<any, IContentTypesState> {
                       {contentType.name}
                     </a>
                     <button
-                      onClick={() =>
-                        this.delete_click(contentType.contentTypeID as Number)
-                      }
+                      onClick={() => this.props.delete(contentType.contentTypeID)}
                     >
                       delete
                     </button>
@@ -38,7 +40,7 @@ class ContentTypes extends React.Component<any, IContentTypesState> {
               }
             )}
         </ul>
-      </Suspense>
+      </>
     );
   }
 }
@@ -47,33 +49,23 @@ const mapStateToProps = (state: IContentTypesState) => state.contentTypes;
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
   delete: (ID: Number) => {
-    fetch(`http://localhost:5000/api/ContentTypes/${ID}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json"
-      }
-    })
-      .then(r => r.json())
-      .then(
-        data => {
-          dispatch(action(DELETE, data.contentTypeID));
-        },
-        e => {
-          console.error(e);
+    if (window.confirm("Delete this?")) {
+      fetch(`http://localhost:5000/api/ContentTypes/${ID}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json"
         }
-      );
+      })
+        .then(response => response.json())
+        .then(response => dispatch(action(CONTENT_TYPE_DELETE, response.contentTypeID)))
+        .catch(exc => dispatch(action(CONTENT_TYPE_ERROR, exc)));
+    }    
   },
   load: () => {
     fetch("http://localhost:5000/api/ContentTypes")
-      .then(r => r.json())
-      .then(
-        data => {
-          dispatch(action(LOAD, data));
-        },
-        e => {
-          console.error(e);
-        }
-      );
+      .then(response => response.json())
+      .then(response => dispatch(action(CONTENT_TYPE_LIST_LOAD, response)))
+      .catch(exc => dispatch(action(CONTENT_TYPE_ERROR, exc)));
   }
 });
